Move more split cell logic around
